+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/render/model/DiskDriveUnbakedModel.java
@@ -18,6 +18,7 @@ import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUti
 
 public class DiskDriveUnbakedModel implements BasicUnbakedModel {
     private static final ResourceLocation BASE_MODEL = createIdentifier("block/disk_drive_base");
+    private static final ResourceLocation DISK_DISCONNECTED_MODEL = createIdentifier("block/disk_disconnected");
     private static final ResourceLocation DISK_MODEL = createIdentifier("block/disk");
 
     @Override
@@ -32,7 +33,8 @@ public class DiskDriveUnbakedModel implements BasicUnbakedModel {
     public @Nullable BakedModel bake(ModelBakery loader, Function<Material, TextureAtlasSprite> textureGetter, ModelState rotationContainer, ResourceLocation modelId) {
         return new DiskDriveBakedModel(
                 loader.bake(BASE_MODEL, rotationContainer),
-                loader.bake(DISK_MODEL, rotationContainer)
+                loader.bake(DISK_MODEL, rotationContainer),
+                loader.bake(DISK_DISCONNECTED_MODEL, rotationContainer)
         );
     }
 }