+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/render/model/baked/DiskDriveBakedModel.java
@@ -3,6 +3,7 @@ package com.refinedmods.refinedstorage2.platform.fabric.render.model.baked;
 import com.refinedmods.refinedstorage2.api.network.node.diskdrive.DiskDriveState;
 import com.refinedmods.refinedstorage2.api.network.node.diskdrive.StorageDiskState;
 import com.refinedmods.refinedstorage2.platform.common.block.BaseBlock;
+import com.refinedmods.refinedstorage2.platform.common.block.entity.diskdrive.DiskDriveBlockEntity;
 import com.refinedmods.refinedstorage2.platform.fabric.render.model.baked.transform.QuadRotator;
 import com.refinedmods.refinedstorage2.platform.fabric.render.model.baked.transform.QuadTranslator;
 
@@ -14,22 +15,49 @@ import net.fabricmc.fabric.api.renderer.v1.render.RenderContext;
 import net.fabricmc.fabric.api.rendering.data.v1.RenderAttachedBlockView;
 import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.core.BlockPos;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.BlockAndTintGetter;
 import net.minecraft.world.level.block.state.BlockState;
 
 public class DiskDriveBakedModel extends ForwardingBakedModel {
     private final BakedModel diskModel;
+    private final BakedModel diskDisconnectedModel;
     private final QuadTranslator[] translators = new QuadTranslator[8];
 
-    public DiskDriveBakedModel(BakedModel baseModel, BakedModel diskModel) {
+    public DiskDriveBakedModel(BakedModel baseModel, BakedModel diskModel, BakedModel diskDisconnectedModel) {
         this.wrapped = baseModel;
         this.diskModel = diskModel;
+        this.diskDisconnectedModel = diskDisconnectedModel;
 
         int i = 0;
         for (int y = 0; y < 4; ++y) {
             for (int x = 0; x < 2; ++x) {
                 translators[i++] = new QuadTranslator(x == 0 ? -(2F / 16F) : -(9F / 16F), -((y * 3F) / 16F) - (2F / 16F), 0);
             }
+        }
+    }
+
+    @Override
+    public boolean isVanillaAdapter() {
+        return false;
+    }
+
+    @Override
+    public void emitItemQuads(ItemStack stack, Supplier<Random> randomSupplier, RenderContext context) {
+        context.fallbackConsumer().accept(wrapped);
+        CompoundTag tag = BlockItem.getBlockEntityData(stack);
+        if (tag == null) {
+            return;
+        }
+        for (int i = 0; i < translators.length; ++i) {
+            if (!DiskDriveBlockEntity.hasDisk(tag, i)) {
+                continue;
+            }
+            context.pushTransform(translators[i]);
+            context.fallbackConsumer().accept(diskDisconnectedModel);
+            context.popTransform();
         }
     }